Create README.md
# illness-treatment

This database has as it's main purpose the disponibility and accesibilty for both doctors and patients on the same hospital. Meaning that
in order to accelerate any hospital process and increase life quality, something like this might be mandatory in our lives.

Something important to realize is that all our entities have many to many relationships, so several additional tables had to be added. 
Everything shoud be performed with the console.